Fix amqp settlement of deliveries and link-credit
This was caused by an upstream breaking API change which we had
inadvertently pulled in as we relied upon the semantic versioning
contract. This commit fixes up the constants to match the new level of
API and also uses an absolute version in the dependencies block rather
than a semver range.

See https://github.com/noodlefrenzy/node-amqp10/issues/315